Update tornado on 04 October 2006
From email:

I dug through the corresponding .csv files, and found the culprit.  The 2005-2006_torn.csv lists several tornadoes that happened in southern Idaho on October 4th 2006.  By comparing to StormData reports - I can see what the problem is:  

The .csv file has a tornado (#948 for the year) that starts at 43.38N/112.75W and ends at 43.49N/113.76W.  But it lists the length as 3 miles.  If it goes over 1 degree of longitude, it's going to be much longer than 3 miles!  And going from 112.75 to 113.76W  would mean the storm would be going WEST over 40 miles - which would be VERY unusual. 

The StormData info for this one is a bit confusing - because it is from two entries because the tornado crossed over a county boundary, but the combined stop/start of the two separate StormData entries would be:  43.41N/112.85W to 43.46N/112.85W.  That's much closer to 3 miles.  There were 4 other tornadoes in southern Idaho that day - and they all seem to correspond to the other entries in the csv and shapefile - so I am pretty confident that the lat/lon on this entry just got messed up somehow.

So, in my mind, if somebody could correct the entry in the .csv and shapefiles for that tornado on 10/04/2006 to have a start/end of 43.41N/112.85W --> 43.46N/112.85W, that would be great!